sync-work-calendar/requirements.txt
google-api-python-client==2.64.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-secret-manager==2.12.5
google_crc32c==1.5.0
Flask==2.2.2